+++ input/input.go
@@ -1,13 +1,11 @@
-package main
+package input
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 )
 
+// model is the model for the input program.
 type model struct{ textinput textinput.Model }
 
 func (m model) Init() tea.Cmd { return textinput.Blink }
@@ -24,19 +22,4 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	var cmd tea.Cmd
 	m.textinput, cmd = m.textinput.Update(msg)
 	return m, cmd
-}
-
-// InputCmd provides a shell script interface for the text input bubble.
-// https://github.com/charmbracelet/bubbles/textinput
-func (pop Pop) InputCmd() {
-	input := textinput.New()
-	input.Focus()
-
-	input.Prompt = pop.Input.Prompt
-	input.Placeholder = pop.Input.Placeholder
-	input.Width = pop.Input.Width
-
-	p := tea.NewProgram(model{input}, tea.WithOutput(os.Stderr))
-	m, _ := p.StartReturningModel()
-	fmt.Println(m.(model).textinput.Value())
 }
