+++ search/command.go
@@ -1,0 +1,9 @@
+package search
+
+import "fmt"
+
+// Run provides a shell script interface for the search bubble.
+// https://github.com/charmbracelet/bubbles/search
+func (o Options) Run() {
+	fmt.Println(o)
+}
