+++ search/search.go
@@ -1,0 +1,1 @@
+package search
